# Use internal SYCL kernel range rounding for RangePolicy ParallelFor