Simplified version of #4405. We need to provide a kernel name for SYCL to do reasonable kernel range rounding internally.