@masterleinad 2021-12-06
Without this, the workgroup size is chosen as a divisor of the workgroup range leading to poor performance if, e.g., the range is a prime number.
For a simple add kernel,
```
n = 524287
AXPBY KK: 8.868529e-01 s 1.321383e+01 GB/s
n = 524288
AXPBY KK: 4.283028e-02 s 2.736090e+02 GB/s
n = 524289
AXPBY KK: 3.133845e-01 s 3.739423e+01 GB/s
```
before this patch
```
n = 524287
AXPBY KK: 4.648326e-02 s 2.521064e+02 GB/s
n = 524288
AXPBY KK: 4.585346e-02 s 2.555696e+02 GB/s
n = 524289
AXPBY KK: 4.766056e-02 s 2.458799e+02 GB/s
```
after this patch.